Update GUI Main.py
Today, I lost my USB backups, so I went backwards on the code that I had created over Friday. I've made it so that the Main Menu, Ordering Tab and view orders tab open up a window. The reason why I had to do this was because the formatting from the Main Menu was being applied to the Ordering Notebook which has different formatting than the Main Menu