Update GUI Main.py
I've fixed my banner so that the buttons work again, when I added the home button the notebook tab values shifted by 1 and I forgot to fix it. I also added a status label in my banner so that the program has a second way of telling the user what page they're on. I started working on the bean ordering tab